requirements.txt
# Dev environment
pip
autopep8

# App
flask
requests
beautifulsoup4
python-dotenv